// frontend/types.ts
export interface Image {
  id: number;
  image_url: string;
  alt_text: string;
  caption?: string;
  show_on_home?: boolean;
  image_type: string;
  linked_object_type: string | null;
}

export interface SocialLink {
  id: number;
  platform: string;
  url: string;
  icon?: string;
  show_on_home?: boolean;
}

export interface Skill {
  id: number;
  name: string;
  slug: string;
  skill_type: string;
  proficiency: string;
  icon?: string;
  show_on_home?: boolean;
}
// ...
export interface WorkExperience {
  id: number;
  company_name: string;
  company_logo: string;
  job_title: string;
  start_date: string;
  end_date?: string;
  is_current: boolean;
  description: string;
  employment_type: string;
  work_mode: string;
  location: string;
  technologies_used: string;
  images?: Image[];
  show_on_home?: boolean;
}

export interface Education {
  id: number;
  institution: string;
  degree: string;
  field_of_study: string;
  start_date: string;
  end_date?: string;
  is_current: boolean;
  logo: string;
  description: string;
  slug: string;
  images?: Image[];
  show_on_home?: boolean;
}

export interface Certificate {
  id: number;
  title: string;
  issuing_organization: string;
  organization_logo?: string;
  issue_date: string;
  credential_url: string;
  credential_id?: string;
  certificate_image: string;
  description?: string;
  skills: string;
  slug: string;
  images?: Image[];
  show_on_home?: boolean;
}

export interface Achievement {
  id: number;
  slug: string;
  title: string;
  achievement_type: string;
  issuer: string;
  date: string;
  description: string;
  url: string;
  image: string;
  images?: Image[];
  show_on_home?: boolean;
}

export interface Testimonial {
  id: number;
  slug: string;
  author_name: string;
  author_title: string;
  author_company: string;
  author_image: string;
  content: string;
  rating: number;
  relationship: string;
  linkedin_url: string;
  date: string;
  images?: Image[];
  show_on_home?: boolean;
}